Python Module - Task 1
Over the past year or so Credit One has seen an increase in the number of customers who have defaulted on loans they have secured from various partners, and Credit One, as their credit scoring service, could risk losing business if the problem is not solved right away. They have enlisted the help of our Data Science team to design and implement a creative, empirically sound solution. It is very important that we all understand from the start that this is not a typical data analytics problem as we have been given full authority to solve this problem with whatever tools and methods we need. As such we've elected to use Python and a few different libraries to do the heavy lifting for us. We'll be using a few main libraries for the bulk of our work, but you should not limit yourselves to only using those - feel free to investigate other libraries if you think they will contribute to the best solution.

Our first need is to define the problem within a data science framework and understand the differences between what we been doing with data analytics and what we're going to be doing in this project with data science. Then, you'll establish a dedicated local programming environment where you can do your work without needing to be in one space or another to have access to the tools you need, before finally starting the analysis and soling this problem. I have attached some historical data that you'll be using for this task so you may use it to focus on understating the problem and getting your environment ready for the task now.